Clan Refactoring Complete
### Axios in, request out
Most of the updated clan routes are just switching request for axios and cleaning up alot of stuff.

### Less is more
Now the clan routes should be even simpler than ever to understand, making them flexible to necessary changes without having to know a whole ton about HTTP requests.

### connectSteam.js
I see some account get routes are actually post requests or just use the request framework. They are next on the chopping block!

+++ routes/views/clans/post/leave.js
@@ -2,22 +2,24 @@ const error = require("../../account/post/error");
 let axios = require('axios');
 
 exports = module.exports = function (req, res) {
-  
-  const membershipId = req.user.data.attributes.clan.membershipId;
-  //Run post to endpoint
-  axios.delete(`${process.env.API_URL}/data/clanMembership/${membershipId}`, {
-    headers: {
-      'Authorization': `Bearer ${req.user.token}`
-    }
-  }).then( () => {
-    
-    // Refreshing user
-    error.userUpdate(req,res,'/clans?flash=leave');
-    
-  }).catch(e => {
-    console.log(e.response);
-    res.redirect(`clans?flash=error`);
-    
-  });
-  
+  if (!req.user.data.attributes.clan) {
+    res.redirect('../clans?flash=error&error=missingData');
+  } else {
+    const membershipId = req.user.data.attributes.clan.membershipId;
+    //Run post to endpoint
+    axios.delete(`${process.env.API_URL}/data/clanMembership/${membershipId}`, {
+      headers: {
+        'Authorization': `Bearer ${req.user.token}`
+      }
+    }).then(() => {
+
+      // Refreshing user
+      error.userUpdate(req, res, '/clans?flash=leave');
+
+    }).catch(e => {
+      console.log(e.response);
+      res.redirect(`../clans?flash=error`);
+
+    });
+  }
 };
